jobs/ingestion/requirements.txt
google-cloud-bigquery==3.20.1
google-cloud-storage==2.16.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.1
